Extract nonblocking subprocess context
