+++ perun/utils/__init__.py
@@ -10,13 +10,9 @@ from typing import (
     Iterable,
     Optional,
     Any,
-    Callable,
-    Iterator,
     TYPE_CHECKING,
 )
-import contextlib
 import os
-import shlex
 import shutil
 import subprocess
 
@@ -161,48 +157,3 @@ def run_external_command(cmd_args: list[str], **subprocess_kwargs: Any) -> int:
     process = subprocess.Popen(cmd_args, **subprocess_kwargs)
     process.wait()
     return process.returncode
-
-
-@contextlib.contextmanager
-def nonblocking_subprocess(
-    command: str,
-    subprocess_kwargs: dict[str, Any],
-    termination: Optional[Callable[..., Any]] = None,
-    termination_kwargs: Optional[dict[str, Any]] = None,
-) -> Iterator[subprocess.Popen[bytes]]:
-    """Runs a non-blocking process in the background using subprocess without shell.
-
-    The process handle is available by using the context manager approach. It is possible to
-    supply custom process termination function (and its arguments) that will be used instead of
-    the subprocess.terminate().
-
-    :param str command: the command to run in the background
-    :param dict subprocess_kwargs: additional arguments for the subprocess Popen
-    :param function termination: the custom termination function or None
-    :param dict termination_kwargs: the arguments for the termination function
-    """
-    # Split process and arguments
-    parsed_cmd = shlex.split(command)
-
-    # Do not allow shell=True
-    if "shell" in subprocess_kwargs:
-        del subprocess_kwargs["shell"]
-
-    # Start the process and do not block it (user can tho)
-    with subprocess.Popen(parsed_cmd, shell=False, **subprocess_kwargs) as proc:
-        try:
-            yield proc
-        except Exception:
-            # Re-raise the encountered exception
-            raise
-        finally:
-            # Don't terminate the process if it has already finished
-            if proc.poll() is None:
-                # Use the default termination if the termination handler is not set
-                if termination is None:
-                    proc.terminate()
-                else:
-                    # Otherwise use the supplied termination function
-                    if termination_kwargs is None:
-                        termination_kwargs = {}
-                    termination(**termination_kwargs)
